- fixed regression in click handlers - cleared up the storing/restore of the list position. TODO: re-position the list is still WIP.

+++ app/src/main/java/com/hardbacknutter/nevertoomanybooks/booklist/BooklistBuilder.java
@@ -849,7 +849,7 @@ public class BooklistBuilder
      *
      * @return Actual list position.
      */
-    public int getListPosition(final int rowId) {
+    public int getListPosition(final long rowId) {
         return mRowStateDAO.getListPosition(rowId);
     }
